# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: TransitIntel API
  version: 0.2.0
  description: TransitIntel — Public transport planning intelligence platform for Provincia di Ancona
servers:
  - url: /api
    description: Base API path
security: []   # Public API — no authentication required
...
paths:
...
  /traffic:
    get:
      operationId: getTraffic
      tags: [traffic]
      summary: Latest traffic snapshots
      parameters:
        - name: from
          in: query
          schema: { type: string }
        - name: to
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 2000, default: 500 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrafficSnapshot"
                  total:
                    type: integer
...
  /gtfs/upload:
    post:
      operationId: uploadGtfs
      tags: [gtfs-upload]
      summary: Upload a GTFS zip file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GtfsUploadResult"
...
  /gtfs/feeds/{id}:
    delete:
      operationId: deleteGtfsFeed
      tags: [gtfs-upload]
      summary: Delete a GTFS feed (cascade)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── GTFS Queries ──────────────────────────────────────────────
...
  /gtfs/analysis:
    get:
      operationId: getGtfsAnalysis
      tags: [gtfs-queries]
      summary: Comprehensive GTFS analysis (frequency, coverage, ranking)
      parameters:
        - name: feedId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GtfsAnalysisResult"

  # ── GTFS Travel ───────────────────────────────────────────────
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        ok: { type: boolean }
        success: { type: boolean }

    # ── Health ─────────────────────────────────────────────────
...
    TrafficSnapshot:
      type: object
      properties:
        id: { type: string }
        segmentId: { type: string }
        lng: { type: number }
        lat: { type: number }
        speed: { type: number }
        freeflowSpeed: { type: number }
        congestionLevel: { type: number }
        capturedAt: { type: string, format: date-time }
...
    GtfsUploadResult:
      type: object
      properties:
        success: { type: boolean }
        feedId: { type: string }
        stopsImported: { type: integer }
        routesImported: { type: integer }
        tripsImported: { type: integer }
        stopTimesImported: { type: integer }
        calendarRows: { type: integer }
        calendarDatesRows: { type: integer }
        shapesImported: { type: integer }
        agencyName: { type: [string, "null"] }
...
    GtfsAnalysisResult:
      type: object
      properties:
        overallScore: { type: integer }
        summary: { type: object }
        frequency: { type: object }
        routeRanking:
          type: array
          items: { type: object }
        poiCoverage: { type: object }
        populationCoverage: { type: object }
        trafficAlignment: { type: object }
        worstServed:
          type: array
          items: { type: object }

    # ── Scenarios ──────────────────────────────────────────────
...